move: Box, Button Component components/ui로 이동 (#10)
