+++ src/components/carousel/TicketCarousel.tsx
@@ -5,8 +5,8 @@ import 'swiper/css/navigation';
 import 'swiper/css/pagination';
 import { Navigation, Pagination } from 'swiper/modules';
 import { PostDetailResponse } from '@api/post/types';
-import { Box as EventInfoBox } from '@components/Box';
-import { Box as DetailInfoBox } from '@components/Box';
+import { Box as EventInfoBox } from '@/components/ui/Box';
+import { Box as DetailInfoBox } from '@/components/ui/Box';
 import { Link, useNavigate } from 'react-router-dom';
 import { FaChevronRight } from 'react-icons/fa6';
 import { handleDateFormat } from '@utils/handleDateFomat';
