+++ src/pages/admin/_AdminPage.tsx
@@ -1,5 +1,5 @@
 import { MenuContainer } from '@components/MenuContainer';
-import { Button } from '@components/Button';
+import { Button } from '@components/ui/Button';
 import { menuList } from './menList';
 import useAuth from '@/hooks/useAuth';
