+++ src/pages/form/upload/_FormUploadPage.tsx
@@ -1,6 +1,6 @@
 import 'react-datepicker/dist/react-datepicker.css';
 import { useEffect, useReducer, useState } from 'react';
-import { Box } from '@components/Box';
+import { Box } from '@/components/ui/Box';
 import { Period } from '@src/types/period-picker/period';
 import { FormUpload, Question } from '@api/form-upload/types';
 import { SubmitButton } from '@components/SubmitButton';
