+++ src/pages/news/upload/_NewsUploadPage.tsx
@@ -1,12 +1,12 @@
 import '@toast-ui/editor/dist/toastui-editor.css';
 import { useEffect, useState } from 'react';
-import { Box as Title } from '@components/Box';
-import { Box as Category } from '@components/Box';
-import { Box as Summary } from '@components/Box';
-import { Box as Desc } from '@components/Box';
-import { Box as FileBox } from '@components/Box';
-import { Box as PeriodBox } from '@components/Box';
-import { Button } from '@components/Button';
+import { Box as Title } from '@/components/ui/Box';
+import { Box as Category } from '@/components/ui/Box';
+import { Box as Summary } from '@/components/ui/Box';
+import { Box as Desc } from '@/components/ui/Box';
+import { Box as FileBox } from '@/components/ui/Box';
+import { Box as PeriodBox } from '@/components/ui/Box';
+import { Button } from '@components/ui/Button';
 import { FaCamera } from 'react-icons/fa6';
 import { useTopBarStore } from '@stores/topBar-stores';
 import { Editor } from '@components/Editor';
