+++ src/pages/post/_PostPage.tsx
@@ -1,8 +1,8 @@
-import { Box } from '@components/Box';
+import { Box } from '@/components/ui/Box';
 import { PostHeader } from '@components/PostHeader';
 import { Carousel } from '@components/carousel/Carousel';
 import parse from 'html-react-parser';
-import { Button } from '@components/Button';
+import { Button } from '@components/ui/Button';
 import { useTopBarStore } from '@stores/topBar-stores';
 import { useGetPostDetail } from '@query-hooks/post';
 import { useLocation, useNavigate } from 'react-router-dom';
