+++ src/pages/profile/FooterButton.tsx
@@ -1,4 +1,4 @@
-import { Button } from '@/components/Button';
+import { Button } from '@components/ui/Button';
 import { useNavigate } from 'react-router-dom';
 
 export const FooterButton = ({
