+++ src/pages/start/_StartPage.tsx
@@ -1,5 +1,5 @@
 import { Link, useNavigate } from 'react-router-dom';
-import { Button } from '@components/Button';
+import { Button } from '@components/ui/Button';
 
 export const StartPage = () => {
   const navigate = useNavigate();
